requirements.txt
boto3
psutil
protobuf>=3.13.0
requests>=2.25.0
requests-aws4auth
python-snappy>=0.5.4
opentelemetry-api
opentelemetry-sdk
opentelemetry-proto
opentelemetry-exporter-prometheus-remote-write
